refactor(corrections): extract triggerBlobDownload utility
Both export.ts and corrections.ts duplicated the blob → object URL →
anchor → click → cleanup dance. Per arch review note on #1155,
extract a shared helper and reuse it from both call sites. Picks up
content-disposition filename when present, falls back to caller-
provided default.
